Installed authentification and file gestion
Installed mongoose-unique-validator, bcrypt and jsonwebtoken for user registration and login.

Created files in "routes", "controllers" and "models" for the users registration and authentification.

Installed multer for file gestion

Udpated the delete function to delete the item file

Created auth middleware

Updated "routes/books" with multer and authentification

Installed sharp and resized images with width and height

+++ backend/controllers/books.js
@@ -1,25 +1,20 @@
 const Book = require('../models/book');
+const fs = require('fs')
 
 exports.createBook = (req, res, next) => {
+    const bookObject = JSON.parse(req.body.book)
+    delete bookObject._id
+    delete bookObject._userId
     const book = new Book({
-        userId: req.body.userId,
-        title: req.body.title,
-        author: req.body.author,
-        imageUrl: req.body.imageUrl,
-        year: req.body.year,
-        genre: req.body.genre,
-        ratings: [
-            {
-                userId: req.body.ratings.userId,
-                grade: req.body.ratings.grade
-            }
-        ],
-        averageRating: req.body.averageRating
-    });
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/resized_${req.file.filename}`
+    })
     book.save()
-        .then(() => { res.status(201).json({ message: 'Post saved successfully!' }) })
-        .catch((error) => { res.status(400).json({ error: error }) });
-}
+        .then(() => { res.status(201).json({ message: 'Book created!' }) })
+        .catch(error => { res.status(400).json({ error }) })
+};
+
 
 exports.getOneBook = (req, res, next) => {
     Book.findOne({
@@ -30,34 +25,62 @@ exports.getOneBook = (req, res, next) => {
 }
 
 exports.modifyBook = (req, res, next) => {
-    const book = new Book({
-        userId: req.body.userId,
-        title: req.body.title,
-        author: req.body.author,
-        imageUrl: req.body.imageUrl,
-        year: req.body.year,
-        genre: req.body.genre,
-        ratings: [
-            {
-                userId: req.body.ratings.userId,
-                grade: req.body.ratings.grade
+    // Verify if there is already a file for this item
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/resized_${req.file.filename}`
+    } : { ...req.body };
+
+
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(403).json({ message: '403 : Unauthorized request' });
+            } else {
+                Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
             }
-        ],
-        averageRating: req.body.averageRating
-    });
-    Book.updateOne({ _id: req.params.id }, book)
-        .then(() => { res.status(201).json({ message: 'Book updated successfully!' }) })
-        .catch((error) => { res.status(400).json({ error: error }) });
-}
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+};
+
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => { res.status(200).json({ message: 'Book deleted!' }) })
-        .catch((error) => { res.status(400).json({ error: error }) });
-}
+    // We search for the item we need to delete
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            // We check if the user is the one who created the item
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                // We delete the file of the selected item
+                fs.unlink(`images/${filename}`, () => {
+                    // Then we delete the item itself
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
+
 
 exports.getAllBooks = (req, res, next) => {
     Book.find()
         .then((books) => { res.status(200).json(books) })
         .catch((error) => { res.status(400).json({ error: error }) });
+}
+
+exports.getBestRating = (req, res, next) => {
+    Book.find().sort({ averageRating: -1 }).limit(3)
+        .then((books) => res.status(200).json(books))
+        .catch((error) => res.status(404).json({ error }));
 }
